[Refactor][Core/Dashboard] Rename snapshot package to api
