+++ python/ray/dashboard/modules/api/component_activities_schema.json
