+++ python/ray/dashboard/modules/api/healthz_agent.py
@@ -3,7 +3,7 @@ from aiohttp.web import Request, Response
 import ray.dashboard.optional_utils as optional_utils
 import ray.dashboard.utils as dashboard_utils
 import ray.exceptions
-from ray.dashboard.modules.snapshot.utils import HealthChecker
+from ray.dashboard.modules.api.utils import HealthChecker
 
 routes = optional_utils.DashboardAgentRouteTable
